+++ src/include/sll/api/string.h
@@ -41,6 +41,19 @@ __API_FUNC_DECL(string_format);
 
 /**
  * \flags api compilation_call func optimizable
+ * \name sll_api_string_flip_case
+ * \group string-api
+ * \desc Docs!
+ * \arg S
+ * \ret Z -> Failure
+ * \ret S -> Success
+ */
+__API_FUNC_DECL(string_flip_case);
+
+
+
+/**
+ * \flags api compilation_call func optimizable
  * \name sll_api_string_replace
  * \group string-api
  * \desc Docs!