+++ src/sll/api/_generated.c
@@ -396,6 +396,25 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_format_raw(const 
 
 
 
+__SLL_API_TYPE_sll_api_string_flip_case sll_api_string_flip_case(__SLL_API_ARGS_sll_api_string_flip_case);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_flip_case_raw(const sll_object_t*const* al,sll_arg_count_t all){
+	if (all<1){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* a=*(al+0);
+	if (!(SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING)){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	sll_string_t out;
+	sll_api_string_flip_case(&(a->dt.s),&out);
+	sll_object_t* out_o=SLL_CREATE();
+	out_o->t=SLL_OBJECT_TYPE_STRING;
+	out_o->dt.s=out;
+	return out_o;
+}
+
+
+
 __SLL_API_TYPE_sll_api_string_replace sll_api_string_replace(__SLL_API_ARGS_sll_api_string_replace);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_replace_raw(const sll_object_t*const* al,sll_arg_count_t all){
 	if (all<1){
@@ -656,6 +675,7 @@ static const internal_function_t _ifunc_data_ptr[]={
 	{"sort_sort",sll_api_sort_sort_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"string_convert",sll_api_string_convert_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"string_format",sll_api_string_format_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
+	{"string_flip_case",sll_api_string_flip_case_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"string_replace",sll_api_string_replace_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"string_to_lower_case",sll_api_string_to_lower_case_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"string_to_title_case",sll_api_string_to_title_case_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
@@ -675,5 +695,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=36;
+const sll_function_index_t _ifunc_size=37;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);